requirements.txt
pytest==8.3.3
pytest-cov==4.1.0
pytest-asyncio==0.24.0
pytest-mock==3.14.0
python-dotenv==1.0.0
openai>=1.0.0
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.3
pydantic==1.10.13  # Using older version that doesn't require Rust
httpx>=0.24.1
pytest-env==1.1.1
